cli: Add `--show-origin` option to `config list`
Add a new `--show-origin` option to `jj config list` to display the
source type and file path (when present) of config values. In cases
where the user config path is a directory, the individual file
containing a given value will be displayed.

The method that captures config file paths,
`config::FileSourceFile::resolve()`, converts them to relative paths.
These may be difficult to recognize when the current working directory
is distant from a config file's location. We therefore attempt to
canonicalize all file paths. Should this fail, we use the path as-is.